Pin deepdiff to later version
Pins deepdiff to the latest version (8) which includes a fix a PyYAML
dependency issue with Cython [1]. Also, replaces the deep diff CLI
command in a functional test with a custom Python script that uses the
DeepDiff class, working around an issue where the deep diff CLI did not
properly recognize numeric values and caused the functional test to fail
spuriously. This commit should fix CI for Python versions 3.10 and 3.11.

[1] https://github.com/seperman/deepdiff/pull/406