Move OpenCV to an optional dependency
Enables a "slim" installation of the package by making OpenCV an
optional dependency. Users can install OpenCV with the new "full"
variant of the package to get support for model fitting. The default
installation will only support forecasting from an existing model. Since
this latter scenario is the most common use case for popcast and
requires a substantially smaller footprint on disk, this default is what
we'll want for environments like the Nextstrain Docker image.